[Refactor] CategoryTree 디렉토리 이동 (#17) - src/utils -> src/datastructures
